+++ Robot/Huvudenhet/src/lcd.c
@@ -114,7 +114,7 @@ void lcd_set_cursor(uint8_t x, uint8_t y) {
 	lcd_cmd(addr+x);
 }
 
-void lcd_distance_sensors(uint8_t *distance_data) {
+void lcd_distance_sensors(uint16_t *distance_data) {
 	lcd_set_cursor(12,0);
 	printf("|%3i", distance_data[0]);
 	lcd_set_cursor(12,1);
